カスタムランキングとR18ランキングのprefetch関数をリファクタリング
- prefetchCustomRanking関数をprefetch.tsに移動
- prefetchR18Ranking関数をprefetch.tsに移動
- 不要なインポートを削除
