+++ src/routes/r18/ranking/{-$type}.tsx
@@ -6,7 +6,7 @@ import {
 } from "@/components/ui/ranking/R18RankingPage";
 import { RankingType, RankingTypeName } from "@/modules/interfaces/RankingType";
 
-import { prefetchR18Ranking } from "@/modules/data/r18ranking";
+import { prefetchR18Ranking } from "@/modules/data/prefetch";
 import {
 	MAIN_PAGE_CACHE_OPTIONS,
 	createCacheHeaders,
